requirements.txt
Package                 Version
----------------------- -----------
anyio                   4.7.0
barcode                 1.0.4
bidict                  0.23.1
blinker                 1.7.0
boltons                 23.0.0
brotlipy                0.7.0
certifi                 2023.5.7
cffi                    1.15.1
chardet                 3.0.4
charset-normalizer      2.0.4
chinesecalendar         1.9.2
click                   8.1.7
conda                   23.5.2
conda-content-trust     0.1.3
conda-libmamba-solver   23.5.0
conda-package-handling  2.1.0
conda_package_streaming 0.8.0
contextlib2             21.6.0
contourpy               1.2.0
cryptography            39.0.1
cycler                  0.12.1
datasketch              1.6.4
dg-logging              0.6.4
distro                  1.9.0
dnspython               2.7.0
email_validator         2.2.0
et-xmlfile              1.1.0
fastapi                 0.115.6
fastapi-cli             0.0.7
Flask                   3.0.3
Flask-SocketIO          5.3.6
fonttools               4.50.0
h11                     0.14.0
httpcore                1.0.7
httptools               0.6.4
httpx                   0.28.1
idna                    2.8
im-web-api-cli          0.1.2
ImageHash               4.3.1
iniconfig               2.0.0
internetdownloadmanager 0.0.2
itsdangerous            2.2.0
Jinja2                  3.1.3
jiter                   0.8.2
joblib                  1.4.2
jsonpatch               1.32
jsonpointer             2.1
kiwisolver              1.4.5
Levenshtein             0.26.1
libmambapy              1.4.1
markdown-it-py          3.0.0
MarkupSafe              2.1.5
marshmallow             3.21.1
matplotlib              3.8.3
mdurl                   0.1.2
numpy                   1.26.4
openai                  1.61.1
opencv-python           4.9.0.80
openpyxl                3.1.5
packaging               23.0
pandas                  2.2.1
pillow                  10.3.0
pip                     25.0
pip-setting             1.0.2
pluggy                  1.4.0
prompt_toolkit          3.0.47
psutil                  6.1.1
pycosat                 0.6.4
pycparser               2.21
pydantic                1.9.2
Pygments                2.18.0
PyMuPDF                 1.25.3
pyOpenSSL               23.0.0
pyparsing               3.1.2
pypeln                  0.4.9
PySocks                 1.7.1
pytesseract             0.3.10
pytest                  8.1.1
python-barcode          0.15.1
python-dateutil         2.9.0.post0
python-dotenv           1.0.1
python-engineio         4.9.0
python-Levenshtein      0.26.1
python-multipart        0.0.20
python-socketio         5.11.2
pytube                  15.0.0
pytz                    2024.1
PyWavelets              1.6.0
PyYAML                  6.0.2
pyzbar                  0.1.9
pyzxing                 1.0.2
RapidFuzz               3.12.1
requests                2.29.0
rich                    13.9.4
rich-toolkit            0.12.0
ruamel.yaml             0.17.21
schema                  0.7.5
scipy                   1.13.0
setuptools              68.2.0
shellingham             1.5.4
simple-websocket        1.0.0
six                     1.16.0
sniffio                 1.3.1
starlette               0.41.3
stopit                  1.1.2
toolz                   0.12.0
tqdm                    4.65.0
typer                   0.15.1
typing_extensions       4.12.2
tzdata                  2024.1
urllib3                 1.25.11
uvicorn                 0.34.0
uvloop                  0.21.0
watchfiles              1.0.3
wcwidth                 0.2.13
websocket-client        1.7.0
websockets              12.0
weibopy                 0.1.2
Werkzeug                3.0.2
wget                    3.2
wheel                   0.41.2
wsproto                 1.2.0
youtube-dl              2021.12.17
zstandard               0.19.0
